Updated server to serve up mockdata for patch requests to certain paths, workign on module.exports and trying to figure out why its not workign as intended.
